+++ __tests__/avatar.test.tsx
@@ -1,19 +1,19 @@
-import { render, screen } from "@testing-library/react";
+import { render, screen } from '@testing-library/react';
 import { NextIntlClientProvider } from 'next-intl';
 import { Avatar, AvatarFallback } from '@components';
 
-describe("Avatar component", () => {
-  it("renders Avatar with default styles", () => {
+describe('Avatar component', () => {
+  it('renders Avatar with default styles', () => {
     render(
       <NextIntlClientProvider locale="en" messages={{}}>
         <Avatar data-testid="avatar" />
       </NextIntlClientProvider>
     );
-    const avatar = screen.getByTestId("avatar");
+    const avatar = screen.getByTestId('avatar');
     expect(avatar).toBeInTheDocument();
   });
 
-  it("renders AvatarFallback with default styles", () => {
+  it('renders AvatarFallback with default styles', () => {
     render(
       <NextIntlClientProvider locale="en" messages={{}}>
         <Avatar>
@@ -21,7 +21,7 @@ describe("Avatar component", () => {
         </Avatar>
       </NextIntlClientProvider>
     );
-    const avatarFallback = screen.getByTestId("avatar-fallback");
+    const avatarFallback = screen.getByTestId('avatar-fallback');
     expect(avatarFallback).toBeInTheDocument();
   });
 });
