+++ src/helpers/index.ts
@@ -1,6 +1,10 @@
-export { decodeBase64, encodeBase64, constructUrl } from './urlHelpers';
+export {
+  decodeBase64,
+  encodeBase64,
+  constructUrl,
+  dateToString,
+} from './requestUtils';
 export { buildRequest } from './buildRequest';
 export { getStatusColor, getMethodColor } from './getStatusColor';
 export { fillVariables } from './fillVariables';
-export { bodyBuilder, dateToString, getFullUrl } from './requestUtils';
 export { sizeGenerator } from './sizeGenerator';
